Tests/LibWeb/Text/expected/css-transform-transition-to-none-invalidation.txt

before x=108, y=80, width=100, height=100
after x=58, y=50, width=100, height=100
